CLASS 3: un-dark the bridge/teasting soundness binaries — the migration wave's dark tests finally RUN
Completes the usize->u64 / board_window / Effect-PQ-variant migration for the SOUNDNESS test binaries the
class-hunt found DARK (docs/CLASS-COMPLETION-HUNT-2026-07-24.md CLASS 3 — a committed migration wave left
these non-compiling, so their teeth couldn't run). Now compiling + PASSING on persvati with
DREGG_ALLOW_UNAUDITED_PQ=1 (the SIGABRT was the dregg-pq assurance gate, not a soundness hole — env-flagged):
- bridge/tests/eth_publics_v2.rs (board_window: None on WholeChainProofBytes): 10 passed, 0 failed.
- teasting/tests/protocol_coverage_gate.rs (explicit Effect::{CreateHybridCell,RotatePqIdentity} +
  StateConstraint::FieldsCountEquals arms + the coverage ratchet 24->25): 3 passed, 0 failed.
- teasting/tests/coverage_state_constraints.rs (index as u64): 27 passed, 0 failed.
- teasting/tests/fuzz_turns.rs (index as u64): 5 passed, 0 failed.

Verified on persvati (native-anchor) off the jammed local lock. The soundness teeth that catch soundness
holes now RUN — the un-darkening the whole CLASS-3 chase was for.

+++ teasting/tests/fuzz_turns.rs
@@ -213,7 +213,7 @@ fn test_fuzz_turns_conservation_and_nonce() {
                     updated: vec![(
                         cell_id,
                         CellStateDelta {
-                            field_updates: vec![(*index, *value)],
+                            field_updates: vec![(*index as u64, *value)],
                             nonce_increment: false,
                             balance_change: 0,
                             permission_changes: None,
@@ -481,7 +481,7 @@ fn test_fuzz_mixed_effects_500_turns() {
                     updated: vec![(
                         cell_id,
                         CellStateDelta {
-                            field_updates: vec![(*index, *value)],
+                            field_updates: vec![(*index as u64, *value)],
                             nonce_increment: false,
                             balance_change: 0,
                             permission_changes: None,
